+++ tablib/compat.py
@@ -25,25 +25,25 @@ if is_py3:
     import tablib.packages.xlwt3 as xlwt
     from tablib.packages import markup3 as markup
     from tablib.packages import openpyxl3 as openpyxl
+    from tablib.packages.odf3 import opendocument, style, text, table
 
+    import csv
+    from io import StringIO
     # py3 mappings
-    ifilter = filter
-    xrange = range
+
     unicode = str
     bytes = bytes
     basestring = str
 
 else:
     from cStringIO import StringIO as BytesIO
+    from cStringIO import StringIO
     import tablib.packages.xlwt as xlwt
     from tablib.packages import markup
     from itertools import ifilter
     from tablib.packages import openpyxl
+    from tablib.packages.odf import opendocument, style, text, table
 
-    # py2 mappings
-    xrange = xrange
+    from tablib.packages import unicodecsv as csv
+
     unicode = unicode
-    bytes = str
-    basestring = basestring
-
-